# Optimize core multiplexer functions