--- a/multidb_test.go
+++ b/multidb_test.go
@@ -649,7 +649,7 @@ func TestMultiDB_MultiNode(t *testing.T) {
 
 func TestMultiDB_MultiTx(t *testing.T) {
 	t.Log("All nodes healthy")
-	mdb, mocks, err := multiTestConnect()
+	mdb, mocks, err := multiTestConnect(defaultTestConns)
 	if err != nil {
 		t.Fatal(err)
 	}