This PR brings some mayor optimizations in the Go routine multiplexers, used for Exec, Query and BeginTx on multiple Nodes and transactions. In order to establish the exact overhead of this package, empty `executor` and `txBeginner` implementations are used in the newly added benchmarks.

Some of the optimizations  in `multiquery.go` were done by:

1. Use a single channel (instead of 2) with `result` a struct, containing both the returned type and error. This eliminated the need of nil checks and `select` blocks.
2. Using a single channel, there is no more need for the `WaitGroup` and a channel closing Go routine.
3. The returned `done` channel is eliminated, this speeds execution in the case of `MultiNode`, which does not care when Go routines are done.
4. `MultiTx` requires all routines to be terminated before the next call. Therefore it now passes a `WaitGroup`. When `WaitGroup` is nil, it is not used.

Initial benchmarks, before optimizations:

````
Benchmark_beginMultiTx-8           19423             58600 ns/op            6976 B/op        205 allocs/op
Benchmark_nodes2Exec-8            542857              2375 ns/op            4080 B/op          8 allocs/op
Benchmark_mtx2Exec-8              460832              2642 ns/op            4080 B/op          8 allocs/op
Benchmark_multiExec-8              35452             33873 ns/op            3888 B/op          6 allocs/op
Benchmark_multiQuery-8             17300             68745 ns/op            2110 B/op          5 allocs/op
Benchmark_multiQueryRow-8          17296             68638 ns/op             210 B/op          3 allocs/op
````

Latest benchmarks include additional benchmarks with the optional `WaitGroup`. Those end by `_wait` suffix :

https://github.com/moapis/multidb/blob/fe9b6b9c4251654565337a2a91fd2bd0c0d0765f/multiquery_test.go#L487-L502

The performance overhead of the `Multi*` variants in this package have dropped between 3% and 35%, depending on the type of function and `WaitGroup` involvement, which adds an additional synchronization step.

`MultiError` has also been refactored to simplify its usage and make the code more readable. Additionally, it now purges nil errors before returning a `MultiError`.